+++ packages/explorer-ui/components/ChainChart/index.tsx
@@ -320,6 +320,12 @@ export const OverviewChart: React.FC<OverviewChartProps> = ({
                 stackId="a"
                 fill={loading ? 'rgba(255, 255, 255, 0.1)' : '#000000'}
               />
+              <Bar
+                isAnimationActive={false}
+                dataKey="worldchain"
+                stackId="a"
+                fill={loading ? 'rgba(255, 255, 255, 0.1)' : '#FFFFFF'}
+              />
             </>
           )}
         </BarChart>