+++ packages/explorer-ui/pages/_app.tsx
@@ -16,7 +16,7 @@ const App = ({ Component, pageProps }) => {
       <GoogleAnalytics gaId={'G-BBC13LQXBD'} />
       <ApolloProvider client={client}>
         <PageWrapper>
-          <Banner />
+          {/* <Banner /> */}
           <Component {...pageProps} />
           <Analytics />
         </PageWrapper>
@@ -25,13 +25,13 @@ const App = ({ Component, pageProps }) => {
   )
 }
 
-const Banner = () => {
-  return (
-    <div className="flex items-center justify-center w-full h-12 p-4 text-white bg-purple-700">
-      We're updating the explorer at the moment, and some data may be
-      inaccurate. Thank you for your patience.
-    </div>
-  )
-}
+// const Banner = () => {
+//   return (
+//     <div className="flex items-center justify-center w-full h-12 p-4 text-white bg-purple-700">
+//       We're updating the explorer at the moment, and some data may be
+//       inaccurate. Thank you for your patience.
+//     </div>
+//   )
+// }
 
 export default App