+++ packages/rest-api/src/routes/bridgeLimitsRoute.ts
@@ -10,7 +10,7 @@ import { normalizeNativeTokenAddress } from '../middleware/normalizeNativeTokenA
 import { checksumAddresses } from '../middleware/checksumAddresses'
 import { validateRouteExists } from '../validations/validateRouteExists'
 
-const router = express.Router()
+const router: express.Router = express.Router()
 
 /**
  * @openapi