+++ packages/widget/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.5](https://github.com/synapsecns/sanguine/compare/@synapsecns/widget@0.9.4...@synapsecns/widget@0.9.5) (2024-11-07)
+
+**Note:** Version bump only for package @synapsecns/widget
+
+
+
+
+
+## [0.9.4](https://github.com/synapsecns/sanguine/compare/@synapsecns/widget@0.9.3...@synapsecns/widget@0.9.4) (2024-11-07)
+
+**Note:** Version bump only for package @synapsecns/widget
+
+
+
+
+
+## [0.9.3](https://github.com/synapsecns/sanguine/compare/@synapsecns/widget@0.9.2...@synapsecns/widget@0.9.3) (2024-10-29)
+
+**Note:** Version bump only for package @synapsecns/widget
+
+
+
+
+
+## [0.9.2](https://github.com/synapsecns/sanguine/compare/@synapsecns/widget@0.9.1...@synapsecns/widget@0.9.2) (2024-10-29)
+
+**Note:** Version bump only for package @synapsecns/widget
+
+
+
+
+
+## [0.9.1](https://github.com/synapsecns/sanguine/compare/@synapsecns/widget@0.9.0...@synapsecns/widget@0.9.1) (2024-10-28)
+
+**Note:** Version bump only for package @synapsecns/widget
+
+
+
+
+
 # [0.9.0](https://github.com/synapsecns/sanguine/compare/@synapsecns/widget@0.8.0...@synapsecns/widget@0.9.0) (2024-10-24)